+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added info tooltip with available template tags to all template settings 
 - Added utility to queue pending sendouts
 - Added sendgrid to webhooks in general settings
+- Added `segmentId` parameter to sendout element query
 
 ### Changed
 - Replaced Frappe charts with ApexCharts